refactor: 스터디, 라운드 서비스 분리
