+++ backend/src/main/java/com/yigongil/backend/domain/round/RoundOfMember.java
@@ -1,4 +1,4 @@
-package com.yigongil.backend.domain.roundofmember;
+package com.yigongil.backend.domain.round;
 
 import com.yigongil.backend.domain.base.BaseEntity;
 import com.yigongil.backend.domain.member.domain.Member;
