+++ backend/src/main/java/com/yigongil/backend/domain/study/Study.java
@@ -4,7 +4,7 @@ import com.yigongil.backend.domain.base.BaseEntity;
 import com.yigongil.backend.domain.meetingdayoftheweek.MeetingDayOfTheWeek;
 import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.round.Round;
-import com.yigongil.backend.domain.roundofmember.RoundOfMember;
+import com.yigongil.backend.domain.round.RoundOfMember;
 import com.yigongil.backend.domain.studymember.Role;
 import com.yigongil.backend.domain.studymember.StudyMember;
 import com.yigongil.backend.domain.studymember.StudyResult;
