+++ backend/src/main/java/com/yigongil/backend/domain/study/StudyController.java
@@ -6,7 +6,6 @@ import com.yigongil.backend.domain.studymember.Role;
 import com.yigongil.backend.request.StudyStartRequest;
 import com.yigongil.backend.request.StudyUpdateRequest;
 import com.yigongil.backend.response.MyStudyResponse;
-import com.yigongil.backend.response.RoundResponse;
 import com.yigongil.backend.response.StudyDetailResponse;
 import com.yigongil.backend.response.StudyListItemResponse;
 import com.yigongil.backend.response.StudyMemberResponse;
@@ -153,15 +152,6 @@ public class StudyController implements StudyApi {
     ) {
         studyService.finish(member, studyId);
         return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/{studyId}/rounds")
-    public ResponseEntity<List<RoundResponse>> findRoundDetailsOfWeek(
-            @PathVariable Long studyId,
-            @RequestParam Integer weekNumber
-    ) {
-        List<RoundResponse> roundResponses = studyService.findRoundDetailsOfWeek(studyId, weekNumber);
-        return ResponseEntity.ok(roundResponses);
     }
 
     @DeleteMapping("/{studyId}/exit")
