+++ backend/src/main/java/com/yigongil/backend/response/MemberOfRoundResponse.java
@@ -1,7 +1,7 @@
 package com.yigongil.backend.response;
 
 import com.yigongil.backend.domain.member.domain.Member;
-import com.yigongil.backend.domain.roundofmember.RoundOfMember;
+import com.yigongil.backend.domain.round.RoundOfMember;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
