+++ backend/src/main/java/com/yigongil/backend/response/MembersCertificationResponse.java
@@ -2,7 +2,7 @@ package com.yigongil.backend.response;
 
 import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.round.Round;
-import com.yigongil.backend.domain.roundofmember.RoundOfMember;
+import com.yigongil.backend.domain.round.RoundOfMember;
 import com.yigongil.backend.exception.NotStudyMemberException;
 import java.util.List;
