+++ backend/src/test/java/com/yigongil/backend/application/MustDoServiceTest.java
@@ -5,10 +5,10 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.mockito.BDDMockito.willReturn;
 
 import com.yigongil.backend.domain.member.domain.Member;
-import com.yigongil.backend.domain.round.MustDoService;
 import com.yigongil.backend.domain.round.Round;
 import com.yigongil.backend.domain.round.RoundRepository;
-import com.yigongil.backend.domain.roundofmember.RoundOfMember;
+import com.yigongil.backend.domain.round.RoundOfMember;
+import com.yigongil.backend.domain.round.RoundService;
 import com.yigongil.backend.exception.InvalidTodoLengthException;
 import com.yigongil.backend.fixture.MemberFixture;
 import com.yigongil.backend.request.MustDoUpdateRequest;
@@ -26,7 +26,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class MustDoServiceTest {
 
     @InjectMocks
-    private MustDoService mustDoService;
+    private RoundService roundService;
 
     @Mock
     private RoundRepository roundRepository;
@@ -61,7 +61,7 @@ class MustDoServiceTest {
 
             //when
             //then
-            assertDoesNotThrow(() -> mustDoService.updateMustDo(member, round.getId(), request));
+            assertDoesNotThrow(() -> roundService.updateMustDo(member, round.getId(), request));
         }
 
         @Test
@@ -73,7 +73,7 @@ class MustDoServiceTest {
 
             //when
             //then
-            assertThatThrownBy(() -> mustDoService.updateMustDo(member, round.getId(), request))
+            assertThatThrownBy(() -> roundService.updateMustDo(member, round.getId(), request))
                     .isInstanceOf(InvalidTodoLengthException.class);
         }
     }
