+++ backend/src/test/java/com/yigongil/backend/fixture/RoundFixture.java
@@ -2,7 +2,7 @@ package com.yigongil.backend.fixture;
 
 import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.round.Round;
-import com.yigongil.backend.domain.roundofmember.RoundOfMember;
+import com.yigongil.backend.domain.round.RoundOfMember;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
